added easyjson, forumCreate fixed

+++ models/forum.go
@@ -1,15 +1,10 @@
 package models
 
-// Информация о форуме.
+//easyjson:json
 type Forum struct {
-	// Название форума.
-	Title string `json:"title"`
-	// Nickname пользователя, который отвечает за форум.
-	User string `json:"user"`
-	// Человекопонятный URL
-	Slug string `json:"slug"`
-	// Общее кол-во сообщений в данном форуме.
-	Posts float32 `json:"posts,omitempty"`
-	// Общее кол-во ветвей обсуждения в данном форуме.
-	Threads float32 `json:"threads,omitempty"`
+	Title   string `json:"title"`
+	User    string `json:"user"`
+	Slug    string `json:"slug"`
+	Posts   int    `json:"posts,omitempty"`
+	Threads int    `json:"threads,omitempty"`
 }
